Struct cr_autophagy_pyo3::Particle

pub struct Particle {
    pub mechanics: Langevin3D,
    pub interaction: TypedInteraction,
}

Cargo or R11 particle depending on the Species of the interaction field.

Fields
mechanics: Langevin3D

The [Langevin3D] motion was chosen to model mechanics of the particle movement

interaction: TypedInteraction

The TypedInteraction assigns the Species to the particle and handles calculation of interactions.