Struct cr_autophagy_pyo3::SimulationSettings

pub struct SimulationSettings {

    pub n_cells_cargo: usize,
    pub n_cells_r11: usize,
    pub cell_radius_cargo: f64,
    pub cell_radius_r11: f64,
    pub mass_cargo: f64,
    pub mass_r11: f64,
    pub damping_cargo: f64,
    pub damping_r11: f64,
    pub kb_temperature_cargo: f64,
    pub kb_temperature_r11: f64,
    pub update_interval: usize,
    pub potential_strength_cargo_cargo: f64,
    pub potential_strength_r11_r11: f64,
    pub potential_strength_cargo_r11: f64,
    pub potential_strength_cargo_r11_avidity: f64,
    pub interaction_range_cargo_cargo: f64,
    pub interaction_range_r11_r11: f64,
    pub interaction_range_r11_cargo: f64,
    pub interaction_relative_neighbour_distance: f64,
    pub dt: f64,
    pub n_times: usize,
    pub save_interval: usize,
    pub extra_saves: Vec<usize>,
    pub n_threads: usize,
    pub domain_size: f64,
    pub domain_cargo_radius_max: f64,
    pub domain_r11_radius_min: f64,
    pub domain_n_voxels: Option<usize>,
    pub storage_name: String,
    pub storage_name_add_date: bool,
    pub show_progressbar: bool,
    pub random_seed: u64,

}

All settings which can be configured by the Python interface.

We aim to provide access to even the more lower-level settings that can influence the results of our simulation. Not all settings do make sense and some combinations can lead to numerical integration problems.

Fields
n_cells_cargo: usize

Number of cargo particles in the simulation.

n_cells_r11: usize

Number of R11 particles in the simulation.

cell_radius_cargo: f64

Radius of the Cargo particles

cell_radius_r11: f64

Radius of the R11 particles

mass_cargo: f64

Mass of the Cargo particles

mass_r11: f64

Mass of the R11 particles

damping_cargo: f64

Damping constant of the Cargo particles of the Langevin mechanics model.

damping_r11: f64

Damping constant of the R11 particles of the Langevin mechanics model.

kb_temperature_cargo: f64

Product of Boltzmann-Constant and temperature of the Langevin mechanics model for Cargo particles.

kb_temperature_r11: f64

Product of Boltzmann-Constant and temperature of the Langevin mechanics model for R11 particles.

update_interval: usize

Product of Boltzmann-Constant and temperature of the Langevin mechanics model for R11 particles.

potential_strength_cargo_cargo: f64

See TypedInteraction

potential_strength_r11_r11: f64

See TypedInteraction

potential_strength_cargo_r11: f64

See TypedInteraction

potential_strength_cargo_r11_avidity: f64

See TypedInteraction

interaction_range_cargo_cargo: f64

See TypedInteraction

interaction_range_r11_r11: f64

See TypedInteraction

interaction_range_r11_cargo: f64

See TypedInteraction

interaction_relative_neighbour_distance: f64

See TypedInteraction

dt: f64

Integration step of the numerical simulation.

n_times: usize

Number of intgration steps done totally.

save_interval: usize

Specifies the frequency at which results are saved as json files. Lower the number for more saved results.

extra_saves: Vec<usize>

Extra iterations at which the simulation should be saved

n_threads: usize

Number of threads to use in the simulation.

domain_size: f64

Overall domain size of the simulation. The lower bound starts at 0.0.

domain_cargo_radius_max: f64

Upper bound of the radius in which the cargo particles will be spawned

domain_r11_radius_min: f64

Minimal radius outside of which r11 particles will be spawned Must be smaller than the domain_size

domain_n_voxels: Option<usize>

See [CartesianCuboid3]

storage_name: String

Name of the folder to store the results in.

storage_name_add_date: bool

Determines if to add the current date at the end of the save path

show_progressbar: bool

Do we want to show a progress bar

random_seed: u64

The seed with which the simulation is initially configured