Struct cr_autophagy_pyo3::TypedInteraction

pub struct TypedInteraction {
    pub species: Species,
    pub cell_radius: f64,
    pub potential_strength_cargo_cargo: f64,
    pub potential_strength_r11_r11: f64,
    pub potential_strength_cargo_r11: f64,
    pub potential_strength_cargo_r11_avidity: f64,
    pub interaction_range_cargo_cargo: f64,
    pub interaction_range_r11_r11: f64,
    pub interaction_range_r11_cargo: f64,
    pub relative_neighbour_distance: f64,
    /* private fields */
}

Interaction potential depending on the other cells species.

Fields
species: Species

Labels particles as Cargo or R11

cell_radius: f64

Radius of the spherical particle

potential_strength_cargo_cargo: f64

Attracting potential strength between individual Cargo particles

potential_strength_r11_r11: f64

Attracting potential strength between individual R11 particles

potential_strength_cargo_r11: f64

Attracting potential strength between individual Cargo and R11 particles

potential_strength_cargo_r11_avidity: f64

Attracting potential strength between multiple Cargo and R11 particles known as avidity. This mechanism only comes into effect when multiple neighbours are present around the two particles.

interaction_range_cargo_cargo: f64

The interaction range (beyond size of the particles) between Cargo particles

interaction_range_r11_r11: f64

The interaction range (beyond size of the particles) between R11 particles

interaction_range_r11_cargo: f64

The interaction range (beyond size of the particles) between Cargo and R11 particles

relative_neighbour_distance: f64

Distance at which other particles of the same species are regarded as neighbours.

Implementations

impl TypedInteraction

pub fn new( species: Species, cell_radius: f64, potential_strength_cargo_cargo: f64, potential_strength_r11_r11: f64, potential_strength_cargo_r11: f64, potential_strength_cargo_r11_avidity: f64, interaction_range_cargo_cargo: f64, interaction_range_r11_r11: f64, interaction_range_r11_cargo: f64, relative_neighbour_distance: f64 ) -> Self

Constructs a new TypedInteraction