Struct cr_autophagy_pyo3::SimulationSettings

pub struct SimulationSettings {

    pub n_cells_cargo: usize,
    pub n_cells_atg11w19: usize,
    pub cell_radius_cargo: f64,
    pub cell_radius_atg11w19: f64,
    pub diffusion_cargo: f64,
    pub diffusion_atg11w19: f64,
    pub temperature_atg11w19: f64,
    pub temperature_cargo: f64,
    pub potential_strength_cargo_cargo: f64,
    pub potential_strength_cargo_atg11w19: f64,
    pub potential_strength_atg11w19_atg11w19: f64,
    pub interaction_range_cargo_cargo: f64,
    pub interaction_range_atg11w19_cargo: f64,
    pub interaction_range_atg11w19_atg11w19: f64,
    pub relative_neighbour_distance: f64,
    pub dt: f64,
    pub t_max: f64,
    pub save_interval: f64,
    pub extra_saves: Vec<usize>,
    pub n_threads: usize,
    pub domain_size: f64,
    pub domain_cargo_radius_max: f64,
    pub domain_atg11w19_radius_min: f64,
    pub domain_n_voxels: Option<usize>,
    pub storage_name: PathBuf,
    pub substitute_date: Option<PathBuf>,
    pub cargo_initials_dir: PathBuf,
    pub show_progressbar: bool,
    pub random_seed: u64,

}

All settings which can be configured by the Python interface.

We aim to provide access to even the more lower-level settings that can influence the results of our simulation. Not all settings do make sense and some combinations can lead to numerical integration problems.

Fields
n_cells_cargo: usize

Number of cargo particles in the simulation.

n_cells_atg11w19: usize

Number of Atg11/19 particles in the simulation.

cell_radius_cargo: f64

Radius of the Cargo particles

cell_radius_atg11w19: f64

Radius of the Atg11/19 particles

diffusion_cargo: f64

Diffusion as given in the [Brownian3D] struct.

diffusion_atg11w19: f64

Diffusion as given in the [Brownian3D] struct.

temperature_atg11w19: f64

Product of Boltzmann-Constant and temperature of the Brownian mechanics model for Atg11/19 particles.

temperature_cargo: f64

Product of Boltzmann-Constant and temperature of the Brownian mechanics model for Atg11/19 particles.

potential_strength_cargo_cargo: f64

See TypedInteraction

potential_strength_cargo_atg11w19: f64

See TypedInteraction

potential_strength_atg11w19_atg11w19: f64

See TypedInteraction

interaction_range_cargo_cargo: f64

See TypedInteraction

interaction_range_atg11w19_cargo: f64

See TypedInteraction

interaction_range_atg11w19_atg11w19: f64

See TypedInteraction

relative_neighbour_distance: f64

See TypedInteraction

dt: f64

Integration step of the numerical simulation.

t_max: f64

Maximum time until which to solve

save_interval: f64

Specifies the frequency at which results are saved as json files. Lower the number for more saved results.

extra_saves: Vec<usize>

Extra iterations at which the simulation should be saved

n_threads: usize

Number of threads to use in the simulation.

domain_size: f64

Overall domain size of the simulation. The lower bound starts at 0.0.

domain_cargo_radius_max: f64

Upper bound of the radius in which the cargo particles will be spawned

domain_atg11w19_radius_min: f64

Minimal radius outside of which atg11w19 particles will be spawned Must be smaller than the domain_size

domain_n_voxels: Option<usize>

See [CartesianCuboid3]

storage_name: PathBuf

Name of the folder to store the results in.

substitute_date: Option<PathBuf>

Possible substitution for the date to store the results

cargo_initials_dir: PathBuf

Path where the cargo initial positions will be stored.

show_progressbar: bool

Do we want to show a progress bar

random_seed: u64

The seed with which the simulation is initially configured

Implementations

impl SimulationSettings

pub fn copy(&self) -> Self

Copies the struct identically

pub fn __copy__(&self) -> Self

Copies the struct identically

pub fn __deepcopy__(&self, _memo: &Bound<'_, SimulationSettings>) -> Self

Copies the struct identically

pub fn __setstate__(&mut self, py: Python<'_>, state: PyObject) -> PyResult<()>

Used to pickle this object

pub fn __getstate__(&self, py: Python<'_>) -> PyResult<PyObject>

Used to pickle this object

pub fn __repr__(&self) -> String

Formats the object

pub fn load_from_file(path: PathBuf) -> PyResult<Self>

Loads the settings from an existing path.

Will fail if the given path is invalid or the given json file does not match memory layout of this object.

pub fn save_to_file(&self, path: PathBuf) -> Result<(), Error>

Saves the object to a file at the given path

pub fn approx_eq(&self, other: &SimulationSettings) -> bool

Compares with other SimulationSettings instance and checks if they are approximately equal.